Struct tracing_systemd::SystemdLayer

pub struct SystemdLayer { /* private fields */ }

A configurable tracing-subscriber layer compatible with journald.

The layer used to format and log events. Can be configured to log to stdout, or directly to journald using the sd-journal feature.

let systemd_layer = SystemdLayer::new()
    .with_target(true)
    .use_level_prefix(false)
    .use_color(true);

Implementations

impl SystemdLayer

pub fn new() -> Self

Creates a new SystemdLayer with default configuration.

pub fn with_thread_ids(self, log_thread_id: bool) -> Self

Sets whether or not to include thread IDs (default: false)

pub fn separate_spans_with(self, span_separator: &'static str) -> Self

Sets the span separator (default: “::”)

pub fn separate_message_with(self, message_separator: &'static str) -> Self

Sets the message separator (default: “: “)

pub fn with_target(self, display_target: bool) -> Self

Sets whether or not to include the target (default: false)

pub fn function_bracket_left(self, function_bracket_left: &'static str) -> Self

Sets the left bracket for function names (default: “(”)

pub fn function_bracket_right( self, function_bracket_right: &'static str ) -> Self

Sets the right bracket for function names (default: “)”)

pub fn arguments_equality(self, arguments_equality: &'static str) -> Self

Sets the equality sign for arguments (default: “: “)

pub fn arguments_separator(self, arguments_separator: &'static str) -> Self

Sets the separator for arguments (default: “, “)

pub fn level_separator(self, level_separator: &'static str) -> Self

Sets the separator for levels (default: “ “)

pub fn thread_id_prefix(self, thread_id_prefix: &'static str) -> Self

Sets the prefix for thread IDs (default: “[”)

pub fn thread_id_suffix(self, thread_id_suffix: &'static str) -> Self

Sets the suffix for thread IDs (default: “] “)

pub fn use_level_prefix(self, use_level_prefix: bool) -> Self

Sets whether or not to use level prefixes (default: true)

pub fn use_color(self, use_color: bool) -> Self

Sets whether or not to use color